Keyboard Navigation inside the popup
Use the cursor keys or hjkl to select cells in the matrix. Press Space
or Return to toggle its color, Shift-Space or Shift-Return to make the
current color explicit.
More vi-like hotkeys are available to save (w) or revert (e/E) changes,
reload (r/R), change scope(i/o, +/-) and toggle turning off uMatrix (q).

To navigate only between the cells that are visible, a NodeIterator is
used. This has the advantage that when additional rows are inserted
(e.g. due to async XHR), NodeIterator automatically picks them up. A
minor disadvantage is that once an iterator has been used up (i.e.
nextNode() returns null), we need to create a new iterator and
fast-forward it to the cell we were previously on (there is sadly no
hasNext()-like functionality).
createNodeIterator()'s acceptNode filter filters on cells by their
common className and uses the node.offsetHeight trick to filter out
those that are invisible due to an ancestor being hidden.